Move lyric editor into individual file.
