+++ osu.Game.Rulesets.Karaoke/Edit/ImportLyric/Header.cs
@@ -7,7 +7,6 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Events;
 using osu.Framework.Screens;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
@@ -156,7 +155,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.ImportLyric
                 {
                     if (enabled)
                         base.SelectTab(tab);
-                });   
+                });
             }
 
             protected override TabItem<IScreen> CreateTabItem(IScreen value) => new HeaderBreadcrumbTabItem(value)
